idk man

+++ MCHexBOT/Protocol/Pakets/Client/Play/SpawnEntityPaket.cs
@@ -11,8 +11,8 @@ namespace MCHexBOT.Pakets.Client.Play
         public double XPosition { get; set; }
         public double YPosition { get; set; }
         public double ZPosition { get; set; }
-        public int Pitch { get; set; }
-        public int Yaw { get; set; }
+        public byte Pitch { get; set; }
+        public byte Yaw { get; set; }
         public int Data { get; set; }
         public short XVelocity { get; set; }
         public short YVelocity { get; set; }
@@ -26,8 +26,8 @@ namespace MCHexBOT.Pakets.Client.Play
             XPosition = minecraftStream.ReadDouble();
             YPosition = minecraftStream.ReadDouble();
             ZPosition = minecraftStream.ReadDouble();
-            Yaw = minecraftStream.ReadByte();
-            Pitch = minecraftStream.ReadByte();
+            Yaw = (byte)minecraftStream.ReadByte();
+            Pitch = (byte)minecraftStream.ReadByte();
             Data = minecraftStream.ReadInt();
             XVelocity = minecraftStream.ReadShort();
             YVelocity = minecraftStream.ReadShort();
@@ -42,8 +42,8 @@ namespace MCHexBOT.Pakets.Client.Play
             minecraftStream.WriteDouble(XPosition);
             minecraftStream.WriteDouble(YPosition);
             minecraftStream.WriteDouble(ZPosition);
-            minecraftStream.WriteByte((byte)Yaw);
-            minecraftStream.WriteByte((byte)Pitch);
+            minecraftStream.WriteByte(Yaw);
+            minecraftStream.WriteByte(Pitch);
             minecraftStream.WriteInt(Data);
             minecraftStream.WriteShort(XVelocity);
             minecraftStream.WriteShort(YVelocity);
